requirements.txt
amqp==5.1.1
APScheduler==3.8.1
asgiref==3.4.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Deprecated==1.2.13
Django==3.2.9
django-celery-beat==2.2.1
django-cors-headers==3.10.0
django-crontab==0.7.1
django-redis==5.2.0
django-rest-swagger==2.2.0
django-sslserver==0.22
django-timezone-field==4.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docutils==0.18.1
drf-yasg==1.20.0
drf-yasg2==1.19.4
et-xmlfile==1.1.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.4
lockfile==0.12.2
luigi==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
numpy==1.22.3
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
Pillow==8.4.0
prompt-toolkit==3.0.24
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.6
python-crontab==2.6.0
python-daemon==2.3.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.3.1
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
schedule==1.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
supervisor==4.2.4
tenacity==6.3.1
tornado==6.1
typing-extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
zipp==3.6.0
pip install django-debug-toolbar==3.4.0